docs-review-agent/requirements.txt
ollama>=0.6.0
pyspellchecker>=0.9.0
requests>=2.32.0
